requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
asyncpg
psycopg2-binary
python-jose
httpx
